requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-MySQLdb==1.0.1
Flask-CORS==4.0.0
Werkzeug==2.3.7
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0

